python: Avoid annoying warning when prefix is the default /usr/local
Python does not support installing modules to /usr/local. If user
install their python modules there, it should be expected that they
won't be found by their python interpreter unless they set PYTHONPATH
manually.

The warning was mostly intended for the case where user installs into
/usr and for some reason the proper install path was not correctly
detected, which was the case on Debian in the past.

Also only print the warning if we actually call one of the methods that
use platlibdir or purelibdir. There is no need to warn if
find_installation() is only used to run custom commands.