make sure no custom_target outputs are named "all" or "meson-internal__*"
Or any other reserved names. We check in add_target that the primary
name of any build target isn't on the forbidden list, but custom_target
allows names that are distinct from the output filenames, so we need to
check those too.

We would eventually still error out all the way at the end, with:

```
ERROR: Multiple producers for Ninja target "all". Please rename your targets.
```

But, if we can check that early and provide the underlying reason
(reserved name) alongside actually useful debugging info (a line
number), then why not?

Refactor the check into a small helper function in the process.