string: Add length() method
While it was possible to access to the last elements of a string via
negative indexing, it was not possible easily iterate through it or to
check nicely whether it was empty or of an expected size (without comparisons
with another expected string).

Also it was not possible to iterate through it, now with the range()
method it would be possible to go char-by-char, but strings doesn't
expose their length so add a method to fill this gap.