interpreter: simplify include_directory extraction
This moves the FeatureNew check for strings out of the helper, and into
the typed_kwargs decorator. This also changes the warning for passing an
absolute path to d_import_dirs to a FeatureDeprecated check. Since we've
hit 1.0, we shouldn't remove deprecated functionality until 2.0, thus
let's treated this as a deprecated "feature", and provide a nicer
message.