modules/qt: look for tools in libexec directory
The 'qt' module fails to detect Qt 6 tools (moc, rcc, and uic)
automatically on Linux and macOS. While qmake is located in bin/ where
detection works correctly, the tools are located in libexec/.

It's possible to help the 'qt' module find the tools by adding Qt's
libexec directory to the PATH environment variable, but this manual
workaround is not ideal.

This patch queries the libexec directory from qmake or pkg-config so
that Qt 6 tools are detected successfully.

Tested with qt6-qtbase-devel-6.2.3-2.fc35.x86_64 on Linux.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>