mtest: shorten test status lines if needed
Test status lines longer than the width of the terminal make the output
in general look ugly and can confuse the new multi-line status lines of
running tests. Shorten such status lines by cutting the name of the test
short so that the line as a whole fits on a single line on the terminal.
Indicate this has happened by appending three periods to the end of the
shortened test name.