mtest: omit exit details when out of terminal width
Omit test exit details from the end of a test status line if there is
not enough width left on the outputting terminal. This avoids the need
to either shift everything left (and break alignment) or continue the
line to a second one (and breaking alignment).

Also unify the status line generating code w.r.t. actual test results
and the "live" status lines of the running tests.