mtest: right align test statuses to save horizontal space
By aligning the test statuses to the right (within the status lines) we
can use the "padding white spaces" to display more of the test's name.

This is important with small terminals as the name of the test get cut
off if the terminal is not wide enough making every character count.