minit: refuse to allow creating broken projects
Some executable names are invalid, and while it's unlikely anyone will
specify such a thing using the --executable argument, it's not unlikely
that people experimenting will attempt to use meson init in a directory
named "test".

This then defaults to that for both the project name and the sample
target name, and the latter produces errors when you try to build it.

Fixes #10321