interpreter: add a strict mode for feature.disable_auto_if
Mesa has a number of cases of long if checks for features to determine
if the feature should be converted from an auto into an enabled or
disabled (using booleans, because it's old). For Mesa's case though, we
want to ensure that the new feature returns is either `auto` or
`disabled`, and never `enabled`, but currently `disable_auto_if` allows
this, only changing auto but ignoring enabled.

Thus I've added a `strict` keyword (false by default), which make it an
error for `disable_auto_if` to return an `enabled()` feature, it will
only return an `auto()` or `disabled()` feature.

I've implemented this in a backwards compatible way, so that existing
users are not affected.

TODO: needs failure tests.