modules: Add a module to put project testing helpers in
We really don't want to expose these as any sort of permanent interface,
it's for our own use, but we do need these sorts of helpers sometimes.

To that end this module can only be used if the `MESON_IS_PROJECT_TEST`
environment variable is set to a !0 integer, otherwise the module
raises an `ImportError`, and the Meson reports "no such module."

The first one is a `catch` method, which allows passing an object, a
method, an expected error message, and then arguments and keyword
arguments to the method to be called and then checks for a Meson
exception.