Add subproject.build_root()
Imagine a C project with a subproject that is its Python bindings. The C
project makes use of the Python bindings to use a higher level language
for writing tests. In order to properly set PYTHONPATH in the tests'
environments, you need to know the subprojects build root, so that the
generated module is visibile when the test imports it. By exposing the
subproject's build root, the parent project can easily set the
PYTHONPATH to resolve the import.