Do not call wine if the WINEPATH is already short enough
We call 'wine cmd' to shorten the wine paths so that we do not exceed
the 2048-char limit for the WINEPATH environment variable.

In the most common case where the path is well below this limit, just
use the unadulterated paths.
This speeds up generating exe_wrapper outputs, and also seems to work
correctly in more cases.

Fixes https://github.com/mesonbuild/meson/issues/10401