mtest: fix rebuilding correct target list before running tests
When manually specifying a target on the CLI, which does not have any
test deps, we calculate that there are no deps to rebuild, we run ninja
without any targets, and this invokes the default "all" rule and maybe
builds a few thousand targets that this specific test does not need.

In such a case we simply should not run ninja at all, since we don't
want to rebuild anything.