linkers: try a bit harder to autodetect the correct linker id
mingw GCC using ld.bfd emits diagnostics that include
"-plugin-opt=-pass-through=-lmoldname" and this triggers a match for
mold when it should not.

Instead, always check the very beginning of the output for the linker
id. This is pretty consistent:

- it is always on stdout

- GCC may put additional things on stderr we don't care about

- swift is bizarre and on some OSes redirects the linker stdout to
  swiftc's stderr, but it will still be the only stderr; we didn't even
  check stderr at all until commit 712cbe056811ebdf0d7358ba07a874717a1c736f

For gold/bfd linkers, the linker id is always the *second* word, after
the legally mandated "GNU" we already check for.