fix invalid type for default_options
We have two checks for the type accepted here. One is the basic
typed_kwargs type-checking, which declares it accepts:

  str | include_directories

The other is the custom validator which further narrows it to strings
with certain option-like properties (needs to be an = assignment).

The former is obviously wrong, which doesn't really matter all that much
but still isn't very nice...

Introduced in commit f34013fb08b8d24d570c96084c5d58c5eaf4f5da.