fix totally invalid import of an import
MachineChoice is a mesonlib object, not a compilers object, so it makes
no sense to import it from the latter simply because the latter imports
it too. This results in brittle module dependencies and everything
breaking when a refactor removes it from the latter.

... also it is a typing-only import so while we are fixing it to import
from the right place, we can also put it in a type-checking block.