always run external projects multi-threaded if possible
The check for if the project supports the -j flag was needlessly
complex. We support two types of project:
- waf, always supports -j
- make, if GNU, supports -j

We never checked waf, and the make check assumed that the entire
command, rather than just the last component, was "make". It also
neglects "gmake".

Since any possible build command *may* support -j, always run the
--version check. Detect either build command in the output.